backend/motion-server/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
pydantic-settings==2.12.0
pytest==9.0.0
opencv-python==4.10.0.84
mediapipe==0.10.21
numpy==1.26.4

# Pytorch with CUDA 12.4 (CUDA 13.0과 호환)
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchaudio==2.5.1+cu124